refactoring: カテゴリのindexページ作成
