--- src/pages/category/[categoryId]/[pageNum].tsx
@@ -1,136 +1,0 @@
-/** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react'
-import {
-  GetStaticProps,
-  GetStaticPropsContext,
-  GetStaticPaths,
-  InferGetStaticPropsType,
-} from 'next'
-import ArticleTitle from '../../../components/atoms/articleTitle/ArticleTitle'
-import Seo from '../../../components/molecules/Seo'
-import PostArchive from '../../../components/organisms/post/PostArchive'
-import BlogLayout from '../../../components/templates/BlogLayout'
-import BlogLayoutBase from '../../../components/templates/BlogLayoutBase'
-import { paginationRange } from '../../../utils/paginationRange'
-import { client } from 'libs/client'
-import Failed from 'src/components/atoms/Failed'
-import { CategoryPagination } from 'src/components/organisms/pagination/CategoryPagination'
-import AsideArchive from 'src/components/templates/aside/AsideArchive'
-import { mediaQuery } from 'src/utils/Breakpoints'
-import { MicrocmsData } from 'types/microcmsData'
-
-const PER_PAGE = 10
-
-
-/**
-* まだパス等の設定はしていません！
-*/
-
-
-export default function CategoryId({
-  blog,
-  totalCount,
-}: InferGetStaticPropsType<typeof getStaticProps>) {
-  if (blog.length === 0) {
-    return <Failed text={'カテゴリに該当する記事はありません。'} />
-  }
-  return (
-    <BlogLayout>
-      <Seo ogpTitle={` ${blog[0].category.name} の記事一覧 | Webのあれこれ`} />
-      <BlogLayoutBase>
-        <ArticleTitle text={`カテゴリ： ${blog[0].category.name} の記事一覧`} />
-        <ul css={postLists}>
-          {blog.map((post: MicrocmsData) => (
-            <PostArchive key={post.id} post={post} /> // 最新ページから取り出した記事
-          ))}
-        </ul>
-
-        <CategoryPagination
-          category={blog[0].category.id}
-          totalCount={totalCount}
-        />
-      </BlogLayoutBase>
-      <AsideArchive />
-    </BlogLayout>
-  )
-}
-
-// SSG: データの取得
-export const getStaticProps: GetStaticProps = async (
-  context: GetStaticPropsContext
-) => {
-  // paramsの型エラー回避のため
-  const params = context?.params?.id
-
-  // paramsの型エラー回避のため
-  if (typeof params === 'undefined')
-    throw new Error('params type is invalid (undefined)')
-  if (typeof params === 'string')
-    throw new Error('params type is invalid (string)')
-
-  // paramsの型エラー回避のため
-  const [categoryId, pageId] = params
-
-  const data = await client.get({
-    endpoint: 'posts',
-    queries: {
-      filters: `category[equals]${categoryId}`,
-      offset: (Number(pageId) - 1) * 10,
-      limit: 10,
-    },
-  })
-
-  return {
-    props: {
-      blog: data.contents,
-      totalCount: data.totalCount,
-    },
-  }
-}
-
-export const getStaticPaths: GetStaticPaths = async () => {
-  // 全てのカテゴリAPI取得
-  const categories = await client.get({ endpoint: 'categories' })
-
-  // pathsに設定する配列の作成
-  const categoryPaths: any = []
-
-  // 各カテゴリ別の記事を取得
-  for (const category of categories.contents) {
-    const { totalCount: totalCountByCategory } = await client.get({
-      endpoint: 'posts',
-      queries: {
-        filters: `category[equals]${category.id}`,
-        limit: 1,
-        fields: 'id',
-      },
-    })
-
-    // 各カテゴリ別の記事をのページ数取得
-    const pages = paginationRange(1, Math.ceil(totalCountByCategory / PER_PAGE))
-
-    // 各カテゴリ別の記事一覧のパスを作成
-    pages.forEach((pageNum) => {
-      categoryPaths.push({
-        params: {
-          id: [category.id, String(pageNum)],
-        },
-      })
-    })
-  }
-
-  return {
-    paths: categoryPaths,
-    fallback: false,
-  }
-}
-
-// css
-const postLists = css`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-  ${mediaQuery[2]} {
-    gap: 10px;
-  }
-`
