+++ types/microCmsPostData.tsx
@@ -1,0 +1,8 @@
+export type microCmsPostData = {
+  id: string
+  createdAt: string
+  updatedAt: string
+  publishedAt: string
+  revisedAt: string
+  name: string
+}
